requirements.txt
boto3==1.4.8
botocore==1.8.2
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
docutils==0.14
falcon==1.3.0
google-api-core==0.1.1
google-auth==1.2.1
google-cloud-bigquery==0.28.0
google-cloud-core==0.28.0
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
idna==2.6
jmespath==0.9.3
pika==0.11.0
protobuf==3.5.0.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
python-dateutil==2.6.1
python-mimeparse==1.6.0
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.11
six==1.11.0
urllib3==1.22
